app/requirements.txt
# FastAPI and web framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.18
jinja2==3.1.4

# Database
sqlalchemy==2.0.43
alembic==1.16.4

# PDF Processing
pypdf==6.0.0
pdfplumber==0.11.7
reportlab==4.0.4

# Data Processing
python-dateutil==2.9.0.post0

# Environment & Configuration
python-dotenv==1.0.1
pydantic-settings==2.7.0